+++ helpers/date/sequence.ts
@@ -5,7 +5,7 @@
  */
 
 import { ensureDate } from './ensureDate';
-import type { DateLike } from './types.model';
+import type { DateLike } from './types';
 
 /**
  * Returns an array of `Date` objects for each day from `start` to `end` (inclusive).
