+++ helpers/date/timestamp.ts
@@ -5,7 +5,7 @@
  */
 
 import { ensureDate } from './ensureDate';
-import type { DateLike } from './types.model';
+import type { DateLike } from './types';
 
 /**
  * Checks if a timestamp is likely in seconds (Java/Unix style) vs milliseconds (JavaScript style)
