+++ helpers/date/weekday.ts
@@ -5,7 +5,7 @@
  */
 
 import { ensureDate } from './ensureDate';
-import type { DateLike } from './types.model';
+import type { DateLike } from './types';
 
 /**
  * Named day-of-week constants following the JavaScript `Date.getDay()`
